requirements.txt
apache-airflow==2.8.1
pandas
scikit-learn
joblib
requests
psycopg2-binary
streamlit
